Conflicts:
	src/libponyc/ast/ast.h
	src/libponyc/ast/error.c
	src/libponyc/ast/lexer.c
	src/libponyc/pkg/package.c
	test/unit/ast_replace.cc
	test/unit/traits.cc
	test/unit/util.h

+++ src/libponyc/ast/builder.c
@@ -68,8 +68,8 @@ typedef struct build_parser_t
   symtab_t* defs;
   builder_ref_t* refs;
   ast_token_id id;
-  int line;
-  int pos;
+  size_t line;
+  size_t pos;
   bool have_token;
   bool had_error;
 } build_parser_t;
